Refactored some methods from SeriesController.cs into SeriesService.cs
