+++ API/Controllers/SeriesController.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using API.Comparators;
 using API.Data;
 using API.Data.Repositories;
 using API.DTOs;
@@ -12,12 +11,10 @@ using API.Entities.Enums;
 using API.Extensions;
 using API.Helpers;
 using API.Services;
-using API.SignalR;
 using Kavita.Common;
 using Kavita.Common.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.Logging;
 
 namespace API.Controllers
@@ -27,16 +24,14 @@ namespace API.Controllers
         private readonly ILogger<SeriesController> _logger;
         private readonly ITaskScheduler _taskScheduler;
         private readonly IUnitOfWork _unitOfWork;
-        private readonly IEventHub _eventHub;
         private readonly ISeriesService _seriesService;
 
 
-        public SeriesController(ILogger<SeriesController> logger, ITaskScheduler taskScheduler, IUnitOfWork unitOfWork, IEventHub eventHub, ISeriesService seriesService)
+        public SeriesController(ILogger<SeriesController> logger, ITaskScheduler taskScheduler, IUnitOfWork unitOfWork, ISeriesService seriesService)
         {
             _logger = logger;
             _taskScheduler = taskScheduler;
             _unitOfWork = unitOfWork;
-            _eventHub = eventHub;
             _seriesService = seriesService;
         }
 
@@ -63,7 +58,7 @@ namespace API.Controllers
         /// <param name="seriesId">Series Id to fetch details for</param>
         /// <returns></returns>
         /// <exception cref="KavitaException">Throws an exception if the series Id does exist</exception>
-        [HttpGet("{seriesId}")]
+        [HttpGet("{seriesId:int}")]
         public async Task<ActionResult<SeriesDto>> GetSeries(int seriesId)
         {
             var userId = await _unitOfWork.UserRepository.GetUserIdByUsernameAsync(User.GetUsername());
@@ -86,22 +81,7 @@ namespace API.Controllers
             var username = User.GetUsername();
             _logger.LogInformation("Series {SeriesId} is being deleted by {UserName}", seriesId, username);
 
-            var series = await _unitOfWork.SeriesRepository.GetSeriesByIdAsync(seriesId);
-
-            var chapterIds = (await _unitOfWork.SeriesRepository.GetChapterIdsForSeriesAsync(new []{seriesId}));
-            var result = await _unitOfWork.SeriesRepository.DeleteSeriesAsync(seriesId);
-
-            if (result)
-            {
-                await _unitOfWork.AppUserProgressRepository.CleanupAbandonedChapters();
-                await _unitOfWork.CollectionTagRepository.RemoveTagsWithoutSeries();
-                await _unitOfWork.CommitAsync();
-                _taskScheduler.CleanupChapters(chapterIds);
-
-                await _eventHub.SendMessageAsync(MessageFactory.SeriesRemoved,
-                    MessageFactory.SeriesRemovedEvent(seriesId, series.Name, series.LibraryId), false);
-            }
-            return Ok(result);
+            return Ok(await _seriesService.DeleteMultipleSeries(new[] {seriesId}));
         }
 
         [Authorize(Policy = "RequireAdminRole")]
@@ -111,25 +91,9 @@ namespace API.Controllers
             var username = User.GetUsername();
             _logger.LogInformation("Series {SeriesId} is being deleted by {UserName}", dto.SeriesIds, username);
 
-            var chapterMappings =
-                await _unitOfWork.SeriesRepository.GetChapterIdWithSeriesIdForSeriesAsync(dto.SeriesIds.ToArray());
+            if (await _seriesService.DeleteMultipleSeries(dto.SeriesIds)) return Ok();
 
-            var allChapterIds = new List<int>();
-            foreach (var mapping in chapterMappings)
-            {
-                allChapterIds.AddRange(mapping.Value);
-            }
-
-            var series = await _unitOfWork.SeriesRepository.GetSeriesByIdsAsync(dto.SeriesIds);
-            _unitOfWork.SeriesRepository.Remove(series);
-
-            if (_unitOfWork.HasChanges() && await _unitOfWork.CommitAsync())
-            {
-                await _unitOfWork.AppUserProgressRepository.CleanupAbandonedChapters();
-                await _unitOfWork.CollectionTagRepository.RemoveTagsWithoutSeries();
-                _taskScheduler.CleanupChapters(allChapterIds.ToArray());
-            }
-            return Ok();
+            return BadRequest("There was an issue deleting the series requested");
         }
 
         /// <summary>
@@ -162,23 +126,7 @@ namespace API.Controllers
         public async Task<ActionResult> UpdateSeriesRating(UpdateSeriesRatingDto updateSeriesRatingDto)
         {
             var user = await _unitOfWork.UserRepository.GetUserByUsernameAsync(User.GetUsername(), AppUserIncludes.Ratings);
-            var userRating = await _unitOfWork.UserRepository.GetUserRatingAsync(updateSeriesRatingDto.SeriesId, user.Id) ??
-                             new AppUserRating();
-
-            userRating.Rating = updateSeriesRatingDto.UserRating;
-            userRating.Review = updateSeriesRatingDto.UserReview;
-            userRating.SeriesId = updateSeriesRatingDto.SeriesId;
-
-            if (userRating.Id == 0)
-            {
-                user.Ratings ??= new List<AppUserRating>();
-                user.Ratings.Add(userRating);
-            }
-
-            _unitOfWork.UserRepository.Update(user);
-
-            if (!await _unitOfWork.CommitAsync()) return BadRequest("There was a critical error.");
-
+            if (!await _seriesService.UpdateSeriesRating(user, updateSeriesRatingDto)) return BadRequest("There was a critical error.");
             return Ok();
         }
 
@@ -323,77 +271,9 @@ namespace API.Controllers
         [HttpPost("metadata")]
         public async Task<ActionResult> UpdateSeriesMetadata(UpdateSeriesMetadataDto updateSeriesMetadataDto)
         {
-            try
+            if (await _seriesService.UpdateSeriesMetadata(updateSeriesMetadataDto))
             {
-                var seriesId = updateSeriesMetadataDto.SeriesMetadata.SeriesId;
-                var series = await _unitOfWork.SeriesRepository.GetSeriesByIdAsync(seriesId);
-                var allTags = (await _unitOfWork.CollectionTagRepository.GetAllTagsAsync()).ToList();
-                if (series.Metadata == null)
-                {
-                    series.Metadata = DbFactory.SeriesMetadata(updateSeriesMetadataDto.Tags
-                        .Select(dto => DbFactory.CollectionTag(dto.Id, dto.Title, dto.Summary, dto.Promoted)).ToList());
-                }
-                else
-                {
-                    series.Metadata.CollectionTags ??= new List<CollectionTag>();
-                    // TODO: Move this merging logic into a reusable code as it can be used for any Tag
-                    var newTags = new List<CollectionTag>();
-
-                    // I want a union of these 2 lists. Return only elements that are in both lists, but the list types are different
-                    var existingTags = series.Metadata.CollectionTags.ToList();
-                    foreach (var existing in existingTags)
-                    {
-                        if (updateSeriesMetadataDto.Tags.SingleOrDefault(t => t.Id == existing.Id) == null)
-                        {
-                            // Remove tag
-                            series.Metadata.CollectionTags.Remove(existing);
-                        }
-                    }
-
-                    // At this point, all tags that aren't in dto have been removed.
-                    foreach (var tag in updateSeriesMetadataDto.Tags)
-                    {
-                        var existingTag = allTags.SingleOrDefault(t => t.Title == tag.Title);
-                        if (existingTag != null)
-                        {
-                            if (series.Metadata.CollectionTags.All(t => t.Title != tag.Title))
-                            {
-                                newTags.Add(existingTag);
-                            }
-                        }
-                        else
-                        {
-                            // Add new tag
-                            newTags.Add(DbFactory.CollectionTag(tag.Id, tag.Title, tag.Summary, tag.Promoted));
-                        }
-                    }
-
-                    foreach (var tag in newTags)
-                    {
-                        series.Metadata.CollectionTags.Add(tag);
-                    }
-                }
-
-                if (!_unitOfWork.HasChanges())
-                {
-                    return Ok("No changes to save");
-                }
-
-                if (await _unitOfWork.CommitAsync())
-                {
-                    foreach (var tag in updateSeriesMetadataDto.Tags)
-                    {
-                        await _eventHub.SendMessageAsync(MessageFactory.SeriesAddedToCollection,
-                            MessageFactory.SeriesAddedToCollectionEvent(tag.Id,
-                                updateSeriesMetadataDto.SeriesMetadata.SeriesId), false);
-                    }
-                    return Ok("Successfully updated");
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "There was an exception when updating metadata");
-                await _unitOfWork.RollbackAsync();
+                return Ok("Successfully updated");
             }
 
             return BadRequest("Could not update metadata");
